requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.5
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==25.3.0
black==24.2.0
certifi==2025.4.26
click==8.1.8
dj-database-url==2.3.0
Django==5.2
djangorestframework==3.16.0
exceptiongroup==1.2.2
flake8==7.0.0
frozenlist==1.6.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.26.0
idna==3.10
iniconfig==2.1.0
isort==5.13.2
magic-filter==1.0.12
mccabe==0.7.0
multidict==6.4.3
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.3.7
pluggy==1.5.0
propcache==0.3.1
psycopg2-binary==2.9.10
pycodestyle==2.11.1
pydantic==2.5.3
pydantic_core==2.14.6
pyflakes==3.2.0
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-django==4.8.0
python-dotenv==1.0.1
python-telegram-bot==20.8
sentry-sdk==1.40.4
sniffio==1.3.1
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.20.0
